requirements.txt
django-admin-tools==0.5.2
django-appconf==1.0.1
django-extra-views==0.6.4
django-haystack==2.4.1
django-imagekit==3.3
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.4.1